% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sii_rii_estimate_sii.R
\name{.sii_rii_estimate_sii}
\alias{.sii_rii_estimate_sii}
\title{intended back-transformation.}
\usage{
.sii_rii_estimate_sii(
  selected,
  validation,
  conf_level = NULL,
  vcov_type = "HC1"
)
}
\description{
The contrast is always explicit: c(-1, 1), meaning:
-1 * Y_disadvantaged + 1 * Y_advantaged.
}
\keyword{internal}
